+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* [#2833](https://github.com/clojure-emacs/cider/pull/2833): Save command history for jack-in with universal arg.
 * [#2828](https://github.com/clojure-emacs/cider/pull/2828): Bind "L" to toggle display of locals during a debug session.
 * [#2800](https://github.com/clojure-emacs/cider/pull/2800): Add support for force-out debugger command.
 * Add support for nREPL 0.8 completion op. It's used if `cider-nrepl` is not available.